new: [internal] Move REST client to new Api controller
