+++ app/Controller/Component/ACLComponent.php
@@ -19,6 +19,12 @@ class ACLComponent extends Component
                 'queryACL' => array(),
                 'restSearch' => array('*'),
             ),
+            'api' => [
+                'rest' => ['perm_auth'],
+                'viewDeprecatedFunctionUse' => [],
+                'openapi' => ['*'],
+                'getApiInfo' => ['*'],
+            ],
             'attributes' => array(
                 'add' => array('perm_add'),
                 'add_attachment' => array('perm_add'),
@@ -505,7 +511,6 @@ class ACLComponent extends Component
                 'eventBlockRule' => array(),
                 'fetchServersForSG' => array('perm_sharing_group'),
                 'filterEventIndex' => array(),
-                'getApiInfo' => array('*'),
                 'getAvailableSyncFilteringRules' => array('*'),
                 'getInstanceUUID' => array('perm_sync'),
                 'getPyMISPVersion' => array('*'),
@@ -531,8 +536,6 @@ class ACLComponent extends Component
                 'releaseUpdateLock' => array(),
                 'resetRemoteAuthKey' => array(),
                 'removeOrphanedCorrelations' => array(),
-                'rest' => array('perm_auth'),
-                'openapi' => array('*'),
                 'restartDeadWorkers' => array(),
                 'restartWorkers' => array(),
                 'serverSettings' => array(),
@@ -549,7 +552,6 @@ class ACLComponent extends Component
                 'updateProgress' => array(),
                 'updateSubmodule' => array(),
                 'uploadFile' => array(),
-                'viewDeprecatedFunctionUse' => array(),
                 'killAllWorkers' => [],
                 'cspReport' => ['*'],
                 'pruneDuplicateUUIDs' => array(),
