+++ app/View/Api/view_deprecated_function_use.ctp
@@ -51,4 +51,3 @@
         }
     }
 
-?>
